Attempt of how leaderboards can work
Create a temp object with the user and their votes in it. This is so we can just sort and add in as needed. Previously we'd have to iterate over a map each time to get to the correct entry but if we use it as a list it should be able to get the direct entry.